refactor(console): rename entrypoints & sharding tags component + path
